// src/app/openapi/model/service.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Service { 
    id?: number;
    subcategoryId?: number;
    businessId?: number;
    title?: string;
    note?: string;
    price?: number;
    avgDuration?: number;
    description?: string;
    template?: boolean;
}